refactor: move runtime configuration handling into separate module
