+++ api/src/main/java/io/javaoperatorsdk/operator/api/ControllerUtils.java
@@ -1,0 +1,35 @@
+package io.javaoperatorsdk.operator.api;
+
+import java.util.Locale;
+
+import io.fabric8.kubernetes.client.CustomResource;
+
+
+public class ControllerUtils {
+    
+    private static final String FINALIZER_NAME_SUFFIX = "/finalizer";
+    
+    public static String getDefaultFinalizerName(String crdName) {
+        return crdName + FINALIZER_NAME_SUFFIX;
+    }
+    
+    public static boolean hasGivenFinalizer(CustomResource resource, String finalizer) {
+        return resource.getMetadata().getFinalizers() != null && resource.getMetadata().getFinalizers().contains(finalizer);
+    }
+    
+    public static String getDefaultNameFor(ResourceController controller) {
+        return getDefaultNameFor(controller.getClass());
+    }
+    
+    public static String getDefaultNameFor(Class<? extends ResourceController> controllerClass) {
+        return getDefaultResourceControllerName(controllerClass.getSimpleName());
+    }
+    
+    public static String getDefaultResourceControllerName(String rcControllerClassName) {
+        final var lastDot = rcControllerClassName.lastIndexOf('.');
+        if (lastDot > 0) {
+            rcControllerClassName = rcControllerClassName.substring(lastDot);
+        }
+        return rcControllerClassName.toLowerCase(Locale.ROOT);
+    }
+}
