+++ pom.xml
@@ -46,6 +46,7 @@
         <module>samples</module>
         <module>quarkus-extension</module>
         <module>api</module>
+        <module>runtime-configuration</module>
     </modules>
