+++ runtime-configuration/src/test/java/io/javaoperatorsdk/operator/config/runtime/ControllerAnnotationProcessorTest.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.processing.annotation;
+package io.javaoperatorsdk.operator.config.runtime;
 
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.CompilationSubject;
