+++ samples/pure-java/src/main/java/io/javaoperatorsdk/operator/sample/PureJavaApplicationRunner.java
@@ -3,12 +3,13 @@ package io.javaoperatorsdk.operator.sample;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.Operator;
+import io.javaoperatorsdk.operator.config.runtime.DefaultConfigurationService;
 
 public class PureJavaApplicationRunner {
 
   public static void main(String[] args) {
     KubernetesClient client = new DefaultKubernetesClient();
-    Operator operator = new Operator(client);
+    Operator operator = new Operator(client, DefaultConfigurationService.instance());
     operator.registerController(new CustomServiceController(client));
   }
 }
