+++ samples/tomcat/src/main/java/io/javaoperatorsdk/operator/sample/TomcatOperator.java
@@ -5,6 +5,8 @@ import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.Operator;
+import io.javaoperatorsdk.operator.config.runtime.DefaultConfigurationService;
+import io.javaoperatorsdk.operator.Operator;
 import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,7 +23,7 @@ public class TomcatOperator {
 
     Config config = new ConfigBuilder().withNamespace(null).build();
     KubernetesClient client = new DefaultKubernetesClient(config);
-    Operator operator = new Operator(client);
+    Operator operator = new Operator(client, DefaultConfigurationService.instance());
 
     TomcatController tomcatController = new TomcatController(client);
     operator.registerControllerForAllNamespaces(tomcatController);
