+++ samples/webserver/src/main/java/io/javaoperatorsdk/operator/sample/WebServerOperator.java
@@ -1,11 +1,13 @@
 package io.javaoperatorsdk.operator.sample;
 
+import java.io.IOException;
+
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.Operator;
-import java.io.IOException;
+import io.javaoperatorsdk.operator.config.runtime.DefaultConfigurationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.takes.facets.fork.FkRegex;
@@ -22,7 +24,7 @@ public class WebServerOperator {
 
     Config config = new ConfigBuilder().withNamespace(null).build();
     KubernetesClient client = new DefaultKubernetesClient(config);
-    Operator operator = new Operator(client);
+    Operator operator = new Operator(client, DefaultConfigurationService.instance());
     operator.registerControllerForAllNamespaces(new WebServerController(client));
 
     new FtBasic(new TkFork(new FkRegex("/health", "ALL GOOD!")), 8080).start(Exit.NEVER);
